Removing the Type facet from the people browser
After discussions with engagement and secretariat we agreed that the
type facet in the people browser didn't make much sense. No one is going
to search by people who are professors. The other issue is that any new
people added to the system for the last 10 months at least don't have a
type, they are just Person. So removing the facet.